bot/requirements.txt
discord.py==1.2.3
google-cloud-storage
